// src/types.ts
export interface GenTemplateType {
  requestFunctionTemplate?: requestFunctionTemplateType;
  topImportPkgTemplate?: topImportPkgTemplateType;
}
// ...
export interface SharedConfig {
  /**
   * 要生成的目标代码类型。
   * 默认为 `typescript`，若设为 `javascript`，会将生成的 `.ts` 文件转换为 `.js` + `.d.ts` 文件并删除原 `.ts` 文件。
   *
   * @default 'typescript'
   */
  target?: 'typescript' | 'javascript';

  /**
   * 是否只生成接口请求内容和返回内容的 TypeSript 类型，是则请求文件和请求函数都不会生成。
   *
   * @default false
   */
  typesOnly?: boolean;

  /**
   * 测试环境名称。
   *
   * **用于获取测试环境域名。**
   *
   * 获取方式：打开项目 --> `设置` --> `环境配置` --> 点开或新增测试环境 --> 复制测试环境名称。
   *
   * @example 'dev'
   */
  devEnvName?: string;

  /**
   * 生产环境名称。
   *
   * **用于获取生产环境域名。**
   *
   * 获取方式：打开项目 --> `设置` --> `环境配置` --> 点开或新增生产环境 --> 复制生产环境名称。
   *
   * @example 'prod'
   */
  prodEnvName?: string;

  /**
   * 输出文件路径。
   *
   * 可以是 `相对路径` 或 `绝对路径`。
   *
   * @example 'src/api/index.ts'
   */
  // outputFilePath?: string | ((interfaceInfo: Interface, changeCase: ChangeCase) => string);
  outputFilePath?: string;

  /**
   * 请求函数文件路径。
   *
   * @default 与 `outputFilePath` 同级目录下的 `request.ts` 文件
   * @example 'src/api/request.ts'
   */
  requestFunctionFilePath?: string;

  /**
   * 如果接口响应的结果是 `JSON` 对象，
   * 且我们想要的数据在该对象下，
   * 那我们就可将 `dataKey` 设为我们想要的数据对应的键。
   *
   * 比如该对象为 `{ code: 0, msg: '成功', data: 100 }`，
   * 我们想要的数据为 `100`，
   * 则我们可将 `dataKey` 设为 `data`。
   *
   * @example 'data'
   */
  dataKey?: string;

  /**
   * 支持生成 React Hooks 代码的相关配置。
   */
  reactHooks?: ReactHooksConfig;

  /**
   * 支持生成 JSON Schema 的相关配置。
   */
  jsonSchema?: JsonSchemaConfig;

  /**
   * 支持生成注释的相关配置。
   */
  comment?: CommentConfig;

  /**
   * 预处理接口信息，返回新的接口信息。可返回 false 排除当前接口。
   *
   * 譬如你想对接口的 `path` 进行某些处理或者想排除某些接口，就可使用该方法。
   *
   * @example
   *
   * ```js
   * interfaceInfo => {
   *   interfaceInfo.path = interfaceInfo.path.replace('v1', 'v2')
   *   return interfaceInfo
   * }
   * ```
   */
  preproccessInterface?(interfaceInfo: Interface, changeCase: ChangeCase): Interface | false;

  /**
   * 获取请求函数的名称。
   *
   * @default changeCase.camelCase(interfaceInfo.parsedPath.name)
   * @param interfaceInfo 接口信息
   * @param changeCase 常用的大小写转换函数集合对象
   * @returns 请求函数的名称
   */
  getRequestFunctionName?(interfaceInfo: ExtendedInterface, changeCase: ChangeCase): string;

  /**
   * 获取请求数据类型的名称。
   *
   * @default changeCase.pascalCase(`${requestFunctionName}Request`)
   * @param interfaceInfo 接口信息
   * @param changeCase 常用的大小写转换函数集合对象
   * @returns 请求数据类型的名称
   */
  getRequestDataTypeName?(interfaceInfo: ExtendedInterface, changeCase: ChangeCase): string;

  /**
   * 获取响应数据类型的名称。
   *
   * @default changeCase.pascalCase(`${requestFunctionName}Response`)
   * @param interfaceInfo 接口信息
   * @param changeCase 常用的大小写转换函数集合对象
   * @returns 响应数据类型的名称
   */
  getResponseDataTypeName?(interfaceInfo: ExtendedInterface, changeCase: ChangeCase): string;
}

/**
 * 分类的配置。
 */
export interface CategoryConfig extends SharedConfig {
  /**
   * 分类 ID，可以设置多个。设为 `0` 时表示全部分类。
   *
   * 如果需要获取全部分类，同时排除指定分类，可以这样：`[0, -20, -21]`，分类 ID 前面的负号表示排除。
   *
   * 获取方式：打开项目 --> 点开分类 --> 复制浏览器地址栏 `/api/cat_` 后面的数字。
   *
   * @example 20
   */
  id: number | number[];
  /**
   * 过滤接口
   */
  filter?: ((path: string) => boolean) | RegExp | string[];
}

/**
 * 项目的配置。
 */
export interface ProjectConfig extends SharedConfig {
  /**
   * 项目的唯一标识。支持多个项目。
   *
   * 获取方式：打开项目 --> `设置` --> `token配置` --> 复制 token。
   *
   * @example 'e02a47122259d0c1973a9ff81cabb30685d64abc72f39edaa1ac6b6a792a647d'
   */
  token?: string;
  /**
   * 设置接口的baseURL
   */
  baseURL?: ((path: string) => string | undefined) | string;

  /**
   * 分类列表。
   */
  categories?: CategoryConfig[];
}

export interface DownloadGitRepoSettings {
  /**
   * 仓库地址
   */
  repository: string;
  /**
   * 分支名
   * @default master
   */
  branch?: string;
  /**
   * 下载目录
   *
   * @default /tmp
   */
  dest?: string;
  /**
   * 全量更新
   * @default false
   * @description 全量更新，忽略git change
   */
  fullUpdate?: boolean;
  /**
   * 筛选
   */
  filter?: ((path: string) => boolean) | RegExp | string[];
  /**
   * 设置接口baseURL
   */
  baseURL?: ((path: string) => string | undefined) | string;
}

/**
 * 服务器的配置。
 */
export interface ServerConfig extends SharedConfig, GenTemplateType {
  /**
   * 服务地址。若服务类型为 `yapi`，此处填其首页地址；若服务类型为 `swagger`，此处填其 json 地址。
   *
   * @example 'http://yapi.foo.bar'
   */
  serverUrl: string;

  /**
   * 服务类型。
   *
   * @default 'swagger'
   */
  serverType?: 'yapi' | 'swagger' | 'git-repo';

  /**
   * 仓库设置
   */
  gitRepoSettings?: DownloadGitRepoSettings;

  /**
   * 项目列表。
   */
  projects?: ProjectConfig | ProjectConfig[];
  /**
   * prettier代码格式化配置文件的路径
   *
   * @default process.cwd()
   */
  prettierConfigPath?: string;
  /**
   * 是否使用默认的请求库
   *
   * @default false
   */
  defaultRequestLib?: boolean;
  /**
   * 代理请求模式，所有请求均请求到指定接口
   */
  proxyInterface?: {
    /**
     * 代理接口
     *
     * @default /admin-interface/proxy/v0/proxy
     */
    path?: string;
  };
}

/** 混合的配置。 */
export type SyntheticalConfig = Partial<
  ServerConfig &
    ProjectConfig &
    CategoryConfig & {
      mockUrl: string;
      devUrl: string;
      prodUrl: string;
    }
>;

/** 配置。 */
export type Config = ServerConfig;